Uploading complete ML model
